// src/interfaces/consultaPoliza.model.ts
export class consultaPolizaModel {
    NumeroPoliza?: string = "";
    CUSPP?: string= "";
    IdTipoDoc?:number = 0;
    TipoDocumento?: string= "0";
    NoDocumetnto?: string= "";
    NoEndoso?:string = "0";
    Paterno?: string= "";
    Materno?: string= "";
    Nombre?: string= "";
    SegundoNombre?: string= "";
    CodPension?:number = 0;
    Fecha?:string= "";
    Error?:string= "";
  }